Update README.md
A straightforward encryption method called the Caesar Cypher swaps out each letter in the plaintext message for a letter located a certain number of letters below in the alphabet. This straightforward yet powerful technique has been used for decades to encrypt communications, and it has a close relationship to cryptography.